Interactions: Move admin/developer related modules to their own namespace and use RequireDeveloperAttribute
